+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - (DEV) Port documents build to mkdocs 1.0.1.
 - Improve the gallery example descriptions.
+- Improve of a few cosmetic steps of the `build gallery-release`.
 
 ### Added
 - Add Minor cosmetic improvements on gallery README.